ORC-1673: Remove test packages `org/apache/orc/tools/[count|merge|sizes]`
