+++ java/tools/src/test/org/apache/orc/tools/TestMergeFiles.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.orc.tools.sizes;
+package org.apache.orc.tools;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
@@ -24,10 +24,12 @@ import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hive.ql.exec.vector.BytesColumnVector;
 import org.apache.hadoop.hive.ql.exec.vector.LongColumnVector;
 import org.apache.hadoop.hive.ql.exec.vector.VectorizedRowBatch;
+import org.apache.orc.CompressionKind;
 import org.apache.orc.OrcFile;
+import org.apache.orc.Reader;
 import org.apache.orc.TypeDescription;
 import org.apache.orc.Writer;
-import org.apache.orc.tools.ColumnSizes;
+import org.apache.orc.tools.MergeFiles;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -38,13 +40,15 @@ import java.nio.file.Paths;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class TestColumnSizes {
+public class TestMergeFiles {
   private Path workDir = new Path(
-      Paths.get(System.getProperty("test.tmp.dir"), "orc-test-sizes").toString());
+      Paths.get(System.getProperty("test.tmp.dir"), "orc-test-merge").toString());
   private Configuration conf;
   private FileSystem fs;
+  private Path testFilePath;
 
   @BeforeEach
   public void openFileSystem() throws Exception {
@@ -53,14 +57,16 @@ public class TestColumnSizes {
     fs.setWorkingDirectory(workDir);
     fs.mkdirs(workDir);
     fs.deleteOnExit(workDir);
+    testFilePath = new Path("TestMergeFiles.testMerge.orc");
+    fs.delete(testFilePath, false);
   }
 
   @Test
-  public void testSizes() throws Exception {
+  public void testMerge() throws Exception {
     TypeDescription schema = TypeDescription.fromString("struct<x:int,y:string>");
     Map<String, Integer> fileToRowCountMap = new LinkedHashMap<>();
-    fileToRowCountMap.put("test-sizes-1.orc", 10000);
-    fileToRowCountMap.put("test-sizes-2.orc", 20000);
+    fileToRowCountMap.put("test-merge-1.orc", 10000);
+    fileToRowCountMap.put("test-merge-2.orc", 20000);
     for (Map.Entry<String, Integer> fileToRowCount : fileToRowCountMap.entrySet()) {
       Writer writer = OrcFile.createWriter(new Path(fileToRowCount.getKey()),
           OrcFile.writerOptions(conf)
@@ -88,13 +94,19 @@ public class TestColumnSizes {
     ByteArrayOutputStream myOut = new ByteArrayOutputStream();
     // replace stdout and run command
     System.setOut(new PrintStream(myOut, false, StandardCharsets.UTF_8));
-    ColumnSizes.main(conf, new String[]{"--summary", workDir.toString()});
+    MergeFiles.main(conf, new String[]{workDir.toString(),
+        "--output", testFilePath.toString()});
     System.out.flush();
     System.setOut(origOut);
     String output = myOut.toString(StandardCharsets.UTF_8);
-    assertTrue(output.contains("Total Files: 2"));
-    assertTrue(output.contains("Total Rows: 30000"));
-    assertTrue(output.contains("Percent  Bytes/Row  Name"));
-    assertTrue(output.contains("_stripe_footer"));
+    System.out.println(output);
+    assertTrue(output.contains("Input files size: 2, Merge files size: 2"));
+
+    try (Reader reader = OrcFile.createReader(testFilePath, OrcFile.readerOptions(conf))) {
+      assertEquals(schema, reader.getSchema());
+      assertEquals(CompressionKind.ZSTD, reader.getCompressionKind());
+      assertEquals(2, reader.getStripes().size());
+      assertEquals(10000 + 20000, reader.getNumberOfRows());
+    }
   }
 }
